astype

+++ data/femnist.py
@@ -88,11 +88,11 @@ class FEMNIST(Dataset):
             self.test_x = self.test_x + test_data[user]['x']
             self.test_y = self.test_y + test_data[user]['y']
         
-        self.train_x = np.array([np.reshape(x, (28, 28).astype('float32')) for x in self.train_x])
-        self.test_x = np.array([np.reshape(x, (28, 28).astype('float32')) for x in self.test_x])
+        self.train_x = np.array([np.reshape(x, (28, 28).astype(float)) for x in self.train_x])
+        self.test_x = np.array([np.reshape(x, (28, 28).astype(float)) for x in self.test_x])
 
-        self.train_y = np.array(self.train_y)
-        self.test_y = np.array(self.test_y)
+        self.train_y = np.array(self.train_y).astype(int)
+        self.test_y = np.array(self.test_y).astype(int)
          
          # get unique label count
         self.unique_labels = list(np.unique(self.train_y))
